fix formatting on save on markdown files adding an empty line after sort-start
markdown formatters like prettier add a new line after a comment.
So `<!-- { sort-start } -->` would always have a new line after it.

So

```
<!-- { sort-start } -->
z
b
a
<!-- { sort-end } -->
```

turns into

```
<!-- { sort-start } -->

a
b
z
<!-- { sort-end } -->
```

There's no way to turn it off afaik so the 'fix' is to add a new line
before the ending sort comment as well (`<!-- { sort-end } -->`).

So now the result is

```
<!-- { sort-start } -->

a
b
z

<!-- { sort-end } -->
```

I would prefer no empty lines at all, but this is the only solution I
could find. This change is only for markdown files.

Other solutions include using the event `onDidSaveTextDocument`
instead, but if you change the text inside of that callback, you, or the
user would have to save again, causing the event to be called again. So
infinite loop and you can never save.

Another solution is to add this extension as a formatter, but I'm not
sure if you can have multiple formatters at once, if you can specify the
order, and I'm too lazy.