fix website bugs & use markdown & use shiki for highlighting instead
the sidebar with the table of contents is kinda jaggedy. It is generated
from the html, but when you change pages first the new html seems to be
added and then the old html is removed, so for like a brief half second
the list has like 50 items and there's a scroll added.

I rerender the table of contents with afterUpdate() which calls it on
*every* component update, onMount() seems to be a bit before the
component fully renders

could add a timeout after onMount(), but there's a possible race
condition there

I've tried other solutions like trying to get the toc rendered during
the markdown to html conversion, and I was successful but I can't figure
out how to escape the markdown container and put it in the sidebar.